requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
sendgrid